requirements.txt
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.4
Flask
Flask-WTF
Flask-Cors
pymongo
openai==0.28
python-dotenv